+++ Sources/xcodeproj/Scheme/XCScheme+LaunchAction.swift
@@ -12,8 +12,8 @@ extension XCScheme {
 
         public enum GPUFrameCaptureMode: String {
             case autoEnabled = "0"
-            case Metal = "1"
-            case OpenGL = "2"
+            case metal = "1"
+            case openGL = "2"
             case disabled = "3"
         }
 
@@ -125,6 +125,7 @@ extension XCScheme {
             super.init(preActions, postActions)
         }
 
+        // swiftlint:disable:next function_body_length
         override init(element: AEXMLElement) throws {
             buildConfiguration = element.attributes["buildConfiguration"] ?? LaunchAction.defaultBuildConfiguration
             selectedDebuggerIdentifier = element.attributes["selectedDebuggerIdentifier"] ?? XCScheme.defaultDebugger
@@ -151,8 +152,10 @@ extension XCScheme {
                 locationScenarioReference = nil
             }
 
-            enableGPUFrameCaptureMode = element.attributes["enableGPUFrameCaptureMode"].flatMap { GPUFrameCaptureMode(rawValue: $0) } ?? LaunchAction.defaultGPUFrameCaptureMode
-            enableGPUValidationMode = element.attributes["enableGPUValidationMode"].flatMap { GPUValidationMode(rawValue: $0) } ?? LaunchAction.defaultGPUValidationMode
+            enableGPUFrameCaptureMode = element.attributes["enableGPUFrameCaptureMode"]
+                .flatMap { GPUFrameCaptureMode(rawValue: $0) } ?? LaunchAction.defaultGPUFrameCaptureMode
+            enableGPUValidationMode = element.attributes["enableGPUValidationMode"]
+                .flatMap { GPUValidationMode(rawValue: $0) } ?? LaunchAction.defaultGPUValidationMode
             enableAddressSanitizer = element.attributes["enableAddressSanitizer"] == "YES"
             enableASanStackUseAfterReturn = element.attributes["enableASanStackUseAfterReturn"] == "YES"
             enableThreadSanitizer = element.attributes["enableThreadSanitizer"] == "YES"