+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Added
 - Added `GPUFrameCaptureMode` and `GPUValidationMode` options to `LaunchAction` https://github.com/tuist/xcodeproj/pull/368 by @schiewe.
 
+### Fixed
+- Fix PBXTarget extension methods https://github.com/tuist/xcodeproj/pull/367 by @danilsmakotin
+
 ## 6.5.0
 
 ### Changed